fix: Move default project settings to project service
